+++ benchmarks/verus/numpy_triple_part_2/files/logic_functions_greater.rs
@@ -8,13 +8,13 @@ verus! {
 // </vc-helpers>
 
 // <vc-spec>
-fn numpy_greater(x1: &Vec<i32>, x2: &Vec<i32>) -> (result: Vec<bool>)
+fn numpy_greater(x1: &Vec<i8>, x2: &Vec<i8>) -> (result: Vec<bool>)
     requires x1.len() == x2.len(),
     ensures 
         result.len() == x1.len(),
         forall|i: int| 0 <= i < result.len() ==> 
-            (result[i] == (x1[i] > x2[i])) &&
-            (result[i] == true ==> !(x2[i] > x1[i])) &&
+            (result[i] == (x1[i] as int > x2[i] as int)) &&
+            (result[i] == true ==> !(x2[i] as int > x1[i] as int)) &&
             (result[i] == true || result[i] == false)
 // </vc-spec>
 // <vc-code>
